+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 ### Prerelease v1.2.3.1 (May 25, 2020)
 
 * Fix bug when in level edit mode that caused update-dnsmasq not loading active config
-
-- i.e. edit service dns forwarding blacklist
+  * i.e. edit service dns forwarding blacklist
 
 ### Prerelease v1.2.3 (May 25, 2020)